+++ BDKSwiftExampleWallet/Extensions/Network+Extensions.swift
@@ -1,0 +1,13 @@
+//
+//  Network+Extensions.swift
+//  BDKSwiftExampleWallet
+//
+//  Created by Matthew Ramsden on 9/4/23.
+//
+
+import Foundation
+import BitcoinDevKit
+
+#if DEBUG
+    let mockKeyClientNetwork = Network.regtest
+#endif
