+++ BDKSwiftExampleWallet/Service/Price Service/PriceService.swift
@@ -34,26 +34,6 @@ extension PriceClient {
 }
 
 #if DEBUG
-    let currentPriceMock = Price(
-        time: 1_693_079_705,
-        usd: 26030,
-        eur: 24508,
-        gbp: 22486,
-        cad: 35314,
-        chf: 23088,
-        aud: 40657,
-        jpy: 3_816_606
-    )
-    let currentPriceMockZero = Price(
-        time: 1_693_079_705,
-        usd: 0,
-        eur: 24508,
-        gbp: 22486,
-        cad: 35314,
-        chf: 23088,
-        aud: 40657,
-        jpy: 3_816_606
-    )
     extension PriceClient {
         static let mock = Self(fetchPrice: { currentPriceMock })
         static let mockPause = Self(fetchPrice: {
